Defines directives clickToggle, copySize, customScrollbar, flip, lettering, resizeFontToFit.
clickToggle - toggles a class on click, optional default handler prevention.

copySize - set's the element's size to that of another.

customScrollbar - initializes mCustomScrollbar on an element.

flip - toggles class flip on a target element.

lettering - enables default letteringjs on an element.

fontToFit -  resizes font width of an element to according to it's content.